requirements.txt
django-admin==3.1.3
sqlite3==3.33.0
apscheduler==3.7.0